move a couple useful cvr report utilities from backend to utils
